requirements.txt
numpy>=1.19.0
opencv-python>=4.5.1
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
mtcnn>=0.1.0
scikit-learn>=0.24.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.61.0
Pillow>=8.2.0
streamlit>=1.0.0
jupyter>=1.0.0
pytest>=6.2.5
